requirements.txt
# ============================================================================
# FLASK & WEB FRAMEWORK
# ============================================================================
Flask==3.0.0
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-WTF==1.2.1

# Flask Dependencies
Jinja2==3.1.2
MarkupSafe==2.1.3
Werkzeug==3.0.1
click==8.1.7
itsdangerous==2.1.2
blinker==1.7.0

# ============================================================================
# SICHERHEIT & AUTHENTICATION
# ============================================================================
argon2-cffi==23.1.0
WTForms==3.1.1

# argon2 Dependencies
cffi==1.16.0
pycparser==2.21

# ============================================================================
# DATENBANK
# ============================================================================
# SQLite ist in Python Standard Library enthalten
# Keine zusätzlichen Packages nötig

# ============================================================================
# CONFIGURATION & ENVIRONMENT
# ============================================================================
python-dotenv==1.0.0

# ============================================================================
# BILDVERARBEITUNG (für path_detector.py)
# ============================================================================
opencv-python==4.8.1.78
numpy==1.26.2

# ============================================================================
# TESTING
# ============================================================================
pytest==7.4.3
pytest-cov==4.1.0

# pytest Dependencies
pluggy==1.3.0
packaging==23.2
iniconfig==2.0.0

# ============================================================================
# DEVELOPMENT (optional - auskommentiert für Produktion)
# ============================================================================
# black==23.12.1              # Code Formatter
# flake8==7.0.0               # Linter
# mypy==1.7.1                 # Type Checker
# pylint==3.0.3               # Code Analyzer